+++ mobileTest/src/components/index.js
@@ -1,0 +1,8 @@
+export { default as CustomButton } from './CustomButton'
+export { default as CustomInput } from './CustomInput'
+export { default as Form } from './Form'
+export { default as GridItem } from './GridItem'
+export { default as IconLink} from './IconLink'
+export { default as ImageChoose} from './ImageChoose'
+export { default as ListItem} from './ListItem'
+export { default as Loading} from './Loading'
